Added in Finalized Audio Clips
I believe this scene is now finished. However, it could be interesting
to implement new skins for the controllers to go with the sci-fi lab
theme. The final bug is that the tracks along the floor do not load
initially, the player must change scenes back to the history hallway and
then they are correctly loaded (texture wise).